Move InventoryItem module to appropriate folder
