+++ cypress/e2e/bulk-edit/in-app/bulk-edit-in-app-former-ids.cy.js
@@ -9,7 +9,7 @@ import FileManager from '../../../support/utils/fileManager';
 import Users from '../../../support/fragments/users/users';
 import generateItemBarcode from '../../../support/utils/generateItemBarcode';
 import BulkEditActions from '../../../support/fragments/bulk-edit/bulk-edit-actions';
-import ItemActions from '../../../support/fragments/inventory/inventoryItem/itemActions';
+import InventoryItems from '../../../support/fragments/inventory/item/inventoryItems';
 
 let user;
 const item = {
@@ -37,7 +37,7 @@ describe('bulk-edit', () => {
 
         cy.getItems({ query: `"barcode"=="${item.itemBarcode}"` }).then((inventoryItem) => {
           inventoryItem.formerIds = [formerId];
-          ItemActions.editItemViaApi(inventoryItem);
+          InventoryItems.editItemViaApi(inventoryItem);
         });
 
         FileManager.createFile(`cypress/fixtures/${validItemFormerIdsFileName}`, formerId);
