+++ cypress/e2e/inventory/item/actions-menu-new-request.cy.js
@@ -10,7 +10,7 @@ import Users from '../../../support/fragments/users/users';
 import InstanceRecordView from '../../../support/fragments/inventory/instanceRecordView';
 import UserEdit from '../../../support/fragments/users/userEdit';
 import ServicePoints from '../../../support/fragments/settings/tenant/servicePoints/servicePoints';
-import ItemActions from '../../../support/fragments/inventory/inventoryItem/itemActions';
+import InventoryItems from '../../../support/fragments/inventory/item/inventoryItems';
 import NewRequest from '../../../support/fragments/requests/newRequest';
 
 describe('inventory', () => {
@@ -198,8 +198,8 @@ describe('inventory', () => {
           'Paged',
         ].forEach((itemStatus) => {
           InventoryInstance.openItemByStatus(itemStatus);
-          ItemActions.openActions();
-          ItemActions.clickNewRequestButton();
+          InventoryItems.openActions();
+          InventoryItems.clickNewRequestButton();
           NewRequest.waitLoadingNewRequestPage();
           NewRequest.checkItemInformationSecton(
             testData.instanceTitle,
@@ -217,8 +217,8 @@ describe('inventory', () => {
             InventorySearchAndFilter.searchByParameter('Title (all)', testData.instanceTitle);
             InstanceRecordView.verifyInstanceRecordViewOpened();
             InventoryInstance.openItemByStatus(itemStatus);
-            ItemActions.openActions();
-            ItemActions.verifyNewRequestButtonIsAbsent();
+            InventoryItems.openActions();
+            InventoryItems.verifyNewRequestButtonIsAbsent();
           },
         );
       },
