+++ cypress/e2e/orders/items-for-receiving-includes-order-closed-statuses.cy.js
@@ -14,7 +14,7 @@ import ItemRecordView from '../../support/fragments/inventory/item/itemRecordVie
 import ServicePoints from '../../support/fragments/settings/tenant/servicePoints/servicePoints';
 import NewLocation from '../../support/fragments/settings/tenant/locations/newLocation';
 import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import ItemActions from '../../support/fragments/inventory/inventoryItem/itemActions';
+import InventoryItems from '../../support/fragments/inventory/item/inventoryItems';
 import ItemRecordEdit from '../../support/fragments/inventory/item/itemRecordEdit';
 import SwitchServicePoint from '../../support/fragments/settings/tenant/servicePoints/switchServicePoint';
 import CheckInActions from '../../support/fragments/check-in-actions/checkInActions';
@@ -87,40 +87,40 @@ describe('orders: Receiving and Check-in', () => {
           // eslint-disable-next-line cypress/no-unnecessary-waiting
           cy.wait(5000);
           InventoryInstance.openItemByBarcodeAndIndex('No barcode');
-          ItemActions.edit();
+          InventoryItems.edit();
           ItemRecordEdit.addBarcode(barcodeForFirstItem);
           ItemRecordEdit.saveAndClose();
           // Need to wait,while instance will be saved
           // eslint-disable-next-line cypress/no-unnecessary-waiting
           cy.wait(5000);
-          ItemActions.closeItem();
+          InventoryItems.closeItem();
           InventoryInstance.openHoldingsAccordion(location.name);
           InventoryInstance.openItemByBarcodeAndIndex('No barcode');
-          ItemActions.edit();
+          InventoryItems.edit();
           ItemRecordEdit.addBarcode(barcodeForSecondItem);
           ItemRecordEdit.saveAndClose();
           // Need to wait,while instance will be saved
           // eslint-disable-next-line cypress/no-unnecessary-waiting
           cy.wait(5000);
-          ItemActions.closeItem();
+          InventoryItems.closeItem();
           InventoryInstance.openHoldingsAccordion(location.name);
           InventoryInstance.openItemByBarcodeAndIndex('No barcode');
-          ItemActions.edit();
+          InventoryItems.edit();
           ItemRecordEdit.addBarcode(barcodeForThirdItem);
           ItemRecordEdit.saveAndClose();
           // Need to wait,while instance will be saved
           // eslint-disable-next-line cypress/no-unnecessary-waiting
           cy.wait(5000);
-          ItemActions.closeItem();
+          InventoryItems.closeItem();
           InventoryInstance.openHoldingsAccordion(location.name);
           InventoryInstance.openItemByBarcodeAndIndex('No barcode');
-          ItemActions.edit();
+          InventoryItems.edit();
           ItemRecordEdit.addBarcode(barcodeForFourItem);
           ItemRecordEdit.saveAndClose();
           // Need to wait,while instance will be saved
           // eslint-disable-next-line cypress/no-unnecessary-waiting
           cy.wait(5000);
-          ItemActions.closeItem();
+          InventoryItems.closeItem();
         });
 
         cy.visit(TopMenu.checkInPath);
@@ -165,7 +165,7 @@ describe('orders: Receiving and Check-in', () => {
         barcodeForFirstItem,
         ITEM_STATUS_NAMES.AVAILABLE,
       );
-      ItemActions.closeItem();
+      InventoryItems.closeItem();
       InventoryInstance.openHoldingsAccordion(location.name);
       InventoryInstance.openItemByBarcodeAndIndex(barcodeForSecondItem);
       ItemRecordView.checkItemDetails(
@@ -173,7 +173,7 @@ describe('orders: Receiving and Check-in', () => {
         barcodeForSecondItem,
         ITEM_STATUS_NAMES.AVAILABLE,
       );
-      ItemActions.closeItem();
+      InventoryItems.closeItem();
       InventoryInstance.openHoldingsAccordion(location.name);
       InventoryInstance.openItemByBarcodeAndIndex(barcodeForFourItem);
       ItemRecordView.checkItemDetails(
@@ -181,7 +181,7 @@ describe('orders: Receiving and Check-in', () => {
         barcodeForFourItem,
         ITEM_STATUS_NAMES.IN_PROCESS,
       );
-      ItemActions.closeItem();
+      InventoryItems.closeItem();
       InventoryInstance.openHoldingsAccordion(location.name);
       InventoryInstance.openItemByBarcodeAndIndex(barcodeForThirdItem);
       ItemRecordView.checkItemDetails(
