+++ cypress/support/fragments/inventory/holdings/inventoryHoldings.js
@@ -1,4 +1,4 @@
-import Items from '../inventoryItem/itemActions';
+import InventoryItems from '../item/inventoryItems';
 
 const getHoldingSources = (searchParams) => cy
   .okapiRequest({
@@ -36,9 +36,11 @@ export default {
     this.getHoldingsRecordsViaApi({ query: `permanentLocationId="${locationId}"` }).then(
       (holdings) => {
         holdings.forEach(({ id: holdingId }) => {
-          Items.getItemViaApi({ query: `holdingsRecordId="${holdingId}"` }).then((items) => {
-            items.forEach(({ id: itemId }) => Items.deleteItemViaApi(itemId));
-          });
+          InventoryItems.getItemViaApi({ query: `holdingsRecordId="${holdingId}"` }).then(
+            (items) => {
+              items.forEach(({ id: itemId }) => InventoryItems.deleteItemViaApi(itemId));
+            },
+          );
           this.deleteHoldingRecordViaApi(holdingId);
         });
       },
