+++ cypress/support/fragments/inventory/item/inventoryItems.js
